requirements.txt
asgiref==3.3.4
Django==3.2
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-modeladmin-reorder==0.3.1
Pillow==8.2.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
